server/requirements.txt
Flask==0.8
PyYAML==3.08
